james-service/requirements/development.txt
alembic==0.9.2
bcrypt==3.1.3
cffi==1.11.5
click==6.7
coverage==4.4.1
Flask==0.12.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.2
Flask-Migrate==2.0.4
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-Testing==0.6.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.6
MarkupSafe==1.0
psycopg2==2.7.1
pycparser==2.18
PyJWT==1.5.0
python-dateutil==2.6.0
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.1.10
Werkzeug==0.12.2
Flask-Mail==0.9.1
flask-marshmallow==0.9.0
marshmallow==2.15.3
Flask-WTF==0.14.2
WTForms==2.2.1
requests==2.9.1
pdfkit==0.6.1
cairocffi==0.9.0
CairoSVG==2.1.3
cssselect2==0.2.1
defusedxml==0.5.0
html5lib==1.0.1
pdfrw==0.4
Pillow==5.2.0
Pyphen==0.9.4
tinycss2==0.6.1
WeasyPrint==0.42.3
webencodings==0.5.1
python-docx==0.8.7
pamqp==1.6.1
rabbitpy==1.0.0